# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduPilot AI - Study Abroad Guidance Chatbot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface University {
  id: number;
  name: string;
  country: string;
  minCgpa: number;
  minIelts: number;
  tuitionEstimateInr: number;
  deadline: string;
  programName: string;
  ranking: number;
}

export type StudentProfileEnglishTest =
  (typeof StudentProfileEnglishTest)[keyof typeof StudentProfileEnglishTest];

export const StudentProfileEnglishTest = {
  IELTS: "IELTS",
  TOEFL: "TOEFL",
  Duolingo: "Duolingo",
  Not_yet: "Not yet",
} as const;

export type StudentProfileBudgetInr =
  (typeof StudentProfileBudgetInr)[keyof typeof StudentProfileBudgetInr];

export const StudentProfileBudgetInr = {
  "10-20L": "10-20L",
  "20-35L": "20-35L",
  "35-50L": "35-50L",
  "50L+": "50L+",
} as const;

export type StudentProfileCountry =
  (typeof StudentProfileCountry)[keyof typeof StudentProfileCountry];

export const StudentProfileCountry = {
  Canada: "Canada",
  USA: "USA",
  UK: "UK",
  Germany: "Germany",
  Australia: "Australia",
} as const;

export interface StudentProfile {
  cgpa: number;
  englishTest: StudentProfileEnglishTest;
  englishScore?: number | null;
  budgetInr: StudentProfileBudgetInr;
  country: StudentProfileCountry;
}

export type UniversityRecommendationCategory =
  (typeof UniversityRecommendationCategory)[keyof typeof UniversityRecommendationCategory];

export const UniversityRecommendationCategory = {
  Safe: "Safe",
  Moderate: "Moderate",
  Ambitious: "Ambitious",
} as const;

export interface UniversityRecommendation {
  university: University;
  category: UniversityRecommendationCategory;
  matchScore: number;
  admissionProbability: number;
}

export interface RecommendationResponse {
  safe: UniversityRecommendation[];
  moderate: UniversityRecommendation[];
  ambitious: UniversityRecommendation[];
  studentProfile: StudentProfile;
  percentageEquivalent: number;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
  studentProfile?: StudentProfile | null;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type ListUniversitiesParams = {
  country?: string;
  minCgpa?: number;
  maxBudgetInr?: number;
};