# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EduPilot AI - Study Abroad Guidance Chatbot API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: openai
    description: OpenAI AI chat, image, and audio operations
  - name: universities
    description: University data operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /universities:
    get:
      operationId: listUniversities
      tags: [universities]
      summary: List all universities
      parameters:
        - name: country
          in: query
          required: false
          schema:
            type: string
        - name: minCgpa
          in: query
          required: false
          schema:
            type: number
        - name: maxBudgetInr
          in: query
          required: false
          schema:
            type: number
      responses:
        "200":
          description: List of universities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/University"
  /universities/recommend:
    post:
      operationId: recommendUniversities
      tags: [universities]
      summary: Get university recommendations based on student profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentProfile"
      responses:
        "200":
          description: University recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationResponse"
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    University:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          type: string
        minCgpa:
          type: number
        minIelts:
          type: number
        tuitionEstimateInr:
          type: number
        deadline:
          type: string
        programName:
          type: string
        ranking:
          type: integer
      required:
        - id
        - name
        - country
        - minCgpa
        - minIelts
        - tuitionEstimateInr
        - deadline
        - programName
        - ranking
    StudentProfile:
      type: object
      properties:
        cgpa:
          type: number
        englishTest:
          type: string
          enum: ["IELTS", "TOEFL", "Duolingo", "Not yet"]
        englishScore:
          type: number
          nullable: true
        budgetInr:
          type: string
          enum: ["10-20L", "20-35L", "35-50L", "50L+"]
        country:
          type: string
          enum: ["Canada", "USA", "UK", "Germany", "Australia"]
      required:
        - cgpa
        - englishTest
        - budgetInr
        - country
    UniversityRecommendation:
      type: object
      properties:
        university:
          $ref: "#/components/schemas/University"
        category:
          type: string
          enum: ["Safe", "Moderate", "Ambitious"]
        matchScore:
          type: integer
        admissionProbability:
          type: integer
      required:
        - university
        - category
        - matchScore
        - admissionProbability
    RecommendationResponse:
      type: object
      properties:
        safe:
          type: array
          items:
            $ref: "#/components/schemas/UniversityRecommendation"
        moderate:
          type: array
          items:
            $ref: "#/components/schemas/UniversityRecommendation"
        ambitious:
          type: array
          items:
            $ref: "#/components/schemas/UniversityRecommendation"
        studentProfile:
          $ref: "#/components/schemas/StudentProfile"
        percentageEquivalent:
          type: number
      required:
        - safe
        - moderate
        - ambitious
        - studentProfile
        - percentageEquivalent
    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
        studentProfile:
          $ref: "#/components/schemas/StudentProfile"
          nullable: true
      required:
        - content
    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error