# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduPilot AI - Study Abroad Guidance Chatbot API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all universities
 */
export const ListUniversitiesQueryParams = zod.object({
  country: zod.coerce.string().optional(),
  minCgpa: zod.coerce.number().optional(),
  maxBudgetInr: zod.coerce.number().optional(),
});

export const ListUniversitiesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  country: zod.string(),
  minCgpa: zod.number(),
  minIelts: zod.number(),
  tuitionEstimateInr: zod.number(),
  deadline: zod.string(),
  programName: zod.string(),
  ranking: zod.number(),
});
export const ListUniversitiesResponse = zod.array(ListUniversitiesResponseItem);

/**
 * @summary Get university recommendations based on student profile
 */
export const RecommendUniversitiesBody = zod.object({
  cgpa: zod.number(),
  englishTest: zod.enum(["IELTS", "TOEFL", "Duolingo", "Not yet"]),
  englishScore: zod.number().nullish(),
  budgetInr: zod.enum(["10-20L", "20-35L", "35-50L", "50L+"]),
  country: zod.enum(["Canada", "USA", "UK", "Germany", "Australia"]),
});

export const RecommendUniversitiesResponse = zod.object({
  safe: zod.array(
    zod.object({
      university: zod.object({
        id: zod.number(),
        name: zod.string(),
        country: zod.string(),
        minCgpa: zod.number(),
        minIelts: zod.number(),
        tuitionEstimateInr: zod.number(),
        deadline: zod.string(),
        programName: zod.string(),
        ranking: zod.number(),
      }),
      category: zod.enum(["Safe", "Moderate", "Ambitious"]),
      matchScore: zod.number(),
      admissionProbability: zod.number(),
    }),
  ),
  moderate: zod.array(
    zod.object({
      university: zod.object({
        id: zod.number(),
        name: zod.string(),
        country: zod.string(),
        minCgpa: zod.number(),
        minIelts: zod.number(),
        tuitionEstimateInr: zod.number(),
        deadline: zod.string(),
        programName: zod.string(),
        ranking: zod.number(),
      }),
      category: zod.enum(["Safe", "Moderate", "Ambitious"]),
      matchScore: zod.number(),
      admissionProbability: zod.number(),
    }),
  ),
  ambitious: zod.array(
    zod.object({
      university: zod.object({
        id: zod.number(),
        name: zod.string(),
        country: zod.string(),
        minCgpa: zod.number(),
        minIelts: zod.number(),
        tuitionEstimateInr: zod.number(),
        deadline: zod.string(),
        programName: zod.string(),
        ranking: zod.number(),
      }),
      category: zod.enum(["Safe", "Moderate", "Ambitious"]),
      matchScore: zod.number(),
      admissionProbability: zod.number(),
    }),
  ),
  studentProfile: zod.object({
    cgpa: zod.number(),
    englishTest: zod.enum(["IELTS", "TOEFL", "Duolingo", "Not yet"]),
    englishScore: zod.number().nullish(),
    budgetInr: zod.enum(["10-20L", "20-35L", "35-50L", "50L+"]),
    country: zod.enum(["Canada", "USA", "UK", "Germany", "Australia"]),
  }),
  percentageEquivalent: zod.number(),
});

/**
 * @summary List all conversations
 */
export const ListOpenaiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
});
export const ListOpenaiConversationsResponse = zod.array(
  ListOpenaiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateOpenaiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListOpenaiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenaiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.date(),
});
export const ListOpenaiMessagesResponse = zod.array(
  ListOpenaiMessagesResponseItem,
);

/**
 * @summary Send a text message and receive a streaming text response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
  studentProfile: zod
    .object({
      cgpa: zod.number(),
      englishTest: zod.enum(["IELTS", "TOEFL", "Duolingo", "Not yet"]),
      englishScore: zod.number().nullish(),
      budgetInr: zod.enum(["10-20L", "20-35L", "35-50L", "50L+"]),
      country: zod.enum(["Canada", "USA", "UK", "Germany", "Australia"]),
    })
    .nullish(),
});