# lib/api-zod/src/generated/types/recommendationResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduPilot AI - Study Abroad Guidance Chatbot API
 * OpenAPI spec version: 0.1.0
 */
import type { StudentProfile } from "./studentProfile";
import type { UniversityRecommendation } from "./universityRecommendation";

export interface RecommendationResponse {
  safe: UniversityRecommendation[];
  moderate: UniversityRecommendation[];
  ambitious: UniversityRecommendation[];
  studentProfile: StudentProfile;
  percentageEquivalent: number;
}