# lib/api-zod/src/generated/types/studentProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduPilot AI - Study Abroad Guidance Chatbot API
 * OpenAPI spec version: 0.1.0
 */
import type { StudentProfileBudgetInr } from "./studentProfileBudgetInr";
import type { StudentProfileCountry } from "./studentProfileCountry";
import type { StudentProfileEnglishTest } from "./studentProfileEnglishTest";

export interface StudentProfile {
  cgpa: number;
  englishTest: StudentProfileEnglishTest;
  englishScore?: number | null;
  budgetInr: StudentProfileBudgetInr;
  country: StudentProfileCountry;
}