# lib/api-zod/src/generated/types/university.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduPilot AI - Study Abroad Guidance Chatbot API
 * OpenAPI spec version: 0.1.0
 */

export interface University {
  id: number;
  name: string;
  country: string;
  minCgpa: number;
  minIelts: number;
  tuitionEstimateInr: number;
  deadline: string;
  programName: string;
  ranking: number;
}