# lib/api-zod/src/generated/types/universityRecommendation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduPilot AI - Study Abroad Guidance Chatbot API
 * OpenAPI spec version: 0.1.0
 */
import type { University } from "./university";
import type { UniversityRecommendationCategory } from "./universityRecommendationCategory";

export interface UniversityRecommendation {
  university: University;
  category: UniversityRecommendationCategory;
  matchScore: number;
  admissionProbability: number;
}